feat: Adds a PaqWait command
This minimally enables scripting usecases. Currently `vim.system` calls
used in various Paq* commands are async and we have to setup autocommand
hooks to wait which is not ideal.

For example, I have an auto update script like this:

    nvim --headless +PaqInstall +PaqWait +qall!

This doesn't work without `+PaqWait` and I have to write a wrapper in
lua.